--- a/src/server/game/Entities/Creature/Creature.cpp
+++ b/src/server/game/Entities/Creature/Creature.cpp
@@ -1108,15 +1108,26 @@ void Creature::Update(uint32 diff)
     }
 
     // For now, do this at the end of the update
-    vis_Update.TUpdate(diff);
-    if (vis_Update.TPassed())
+    // For now, do this at the end of the update 
+    float baseLineDiff = 300.0f; // A normal diff for an active server
+    float scaleFactor = std::min(std::max(p_time / baseLineDiff, 1.0f), 5.0f); // Diff scale 1x->5x
+    uint32 scaledPeriod = GetMap()->GetVisibilityNotifyPeriod() * scaleFactor; 
+    uint32 currentTime = GameTime::GetGameTimeMS();
+    uint32 currentOffset = currentTime % scaledPeriod;
+    uint32 lastOffset = (currentTime - p_time) % scaledPeriod;
+    uint32 guidOffset = GetGUID().GetCounter() % scaledPeriod;
+    // Check if guidOffset was crossed during this frame
+    bool crossed = (lastOffset < currentOffset) ?
+        (guidOffset > lastOffset && guidOffset <= currentOffset) :
+        (guidOffset > lastOffset || guidOffset <= currentOffset);
+    if (crossed)
     {
-        vis_Update.TReset(diff, GetMap()->GetVisibilityNotifyPeriod());
-
         if (isNeedNotify(NOTIFY_VISIBILITY_CHANGED))
         {
             CreatureRelocationNotifier relocate(*this);
-            Cell::VisitAllObjects(this, relocate, 100, false);
+            // Scale range from full down to half based on scaleFactor (1x->5x becomes 1.0->0.5)
+            float rangeScale = 1.0f - ((scaleFactor - 1.0f) / 8.0f);
+            Cell::VisitAllObjects(this, relocate, GetMap()->GetVisibilityRange() * rangeScale, false);
         }
 
         ResetAllNotifies();