--- a/src/server/game/Entities/Player/Player.cpp
+++ b/src/server/game/Entities/Player/Player.cpp
@@ -1420,7 +1420,6 @@ void Player::Update(uint32 p_time)
         (guidOffset > lastOffset || guidOffset <= currentOffset);
     if (crossed)
     {
-        TC_LOG_DEBUG("testnotify", "SendNotify {}", currentTime);
         WorldObject const* viewPoint = m_seer;
         if (viewPoint->isNeedNotify(NOTIFY_VISIBILITY_CHANGED) && (this == viewPoint || viewPoint->IsPositionValid()))
         {
@@ -1433,10 +1432,6 @@ void Player::Update(uint32 p_time)
         }
 
         ResetAllNotifies();
-    }
-    else
-    {
-        TC_LOG_DEBUG("testnotify", "NOT SendNotify {}", currentTime);
     }
 }