Performance improvements for High diffs:

We will scale back the relocation notifies in two ways:

The default notify period is 1 second.  We use a baseline diff of 300ms.  We use a max threshold of 1500ms.
We get a scale factor of 1->5 as the diff scales from 300ms->1500ms, clamped.  meaning for diffs >= 1500 we will
have a notify period of 5 seconds, ensuring we do not send every single notify on every single frame.  If the diff > 5 seconds
this cannot be avoided and will fall off a cliff.

We also use this same diff range to scale back the notify distance from 100 yards -> 50 yards.

To ensure better distribution of notifies we switch to using the guid to determine whether an object will update - this should reduce the clustering of update timers and provide a more consistent distribution.